+++ myems-normalization/meter.py
@@ -399,12 +399,14 @@ def worker(meter):
     ####################################################################################################################
     # Step 4: Insert into energy database
     ####################################################################################################################
-    if len(normalized_values) > 0:
+    while len(normalized_values) > 0:
+        insert_100 = normalized_values[:100]
+        normalized_values = normalized_values[100:]
         try:
             add_values = (" INSERT INTO tbl_meter_hourly (meter_id, start_datetime_utc, actual_value) "
                           " VALUES  ")
 
-            for meta_data in normalized_values:
+            for meta_data in insert_100:
                 add_values += " (" + str(meter['id']) + ","
                 add_values += "'" + meta_data['start_datetime_utc'].isoformat()[0:19] + "',"
                 add_values += str(meta_data['actual_value']) + "), "